ppx-align/pyproject.toml
[project]
name = "ppx-align"
version = "0.1.0"
description = "Alignment, layout tree, and web server for the ppx OCR pipeline"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "typer>=0.12",
    "rich>=14.0",
    "pandas>=2.0",
    "fastparquet>=2025.12.0",
    "pandera>=0.29.0",
    "pillow>=10.0",
    "thefuzz>=0.22.1",
    "nltk>=3.9.3",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "sentence-transformers>=5.3.0",
    "scipy>=1.13",
    "markdown-it-py[plugins]>=4.0.0",
    "linkify-it-py>=2.1.0",
    "mdit-py-plugins>=0.5.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[project.scripts]
ppx-align = "ppx_align.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ppx_align"]
